requirements.txt
google-adk>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.2.0
numpy>=2.0.0
httpx>=0.27.0
backtrader>=1.9.78.123